#1964 Added an SGMContinuousScanController: - Moved the SGM specific logic from AMContinuousScanActionController to SGMContinuousScanController, to remove the CLS dependency from an AM class
