+++ source/acquaman/AMContinuousScanActionController.h
@@ -13,23 +13,14 @@
 
 class AMScanActionControllerScanAssembler;
 class AMListAction3;
-class AMDSClientDataRequest;
-
-class CLSAMDSScalerChannelDetector;
-class CLSAmptekSDD123DetectorNew;
 
 Q_DECLARE_METATYPE(AMScanActionControllerBasicFileWriter::FileWriterError)
 
-#define AMCONTINUOUSSCANACTIONCONTROLLER_COULD_NOT_ADD_DETECTOR 285000
-#define AMCONTINUOUSSCANACTIONCONTROLLER_REQUIRED_DATA_MISSING  285001
-#define AMCONTINUOUSSCANACTIONCONTROLLER_INITIAL_ENCODER_POSITION_MISSING  285002
-#define AMCONTINUOUSSCANACTIONCONTROLLER_BAD_SCALER_DATAHOLDER_TYPE  285003
-#define AMCONTINUOUSSCANACTIONCONTROLLER_SCALER_CHANNEL_MISMATCH  285004
-
-#define AMCONTINUOUSSCANACTIONCONTROLLER_FILE_ALREADY_EXISTS 285006
-#define AMCONTINUOUSSCANACTIONCONTROLLER_COULD_NOT_OPEN_FILE 285007
-#define AMCONTINUOUSSCANACTIONCONTROLLER_UNKNOWN_FILE_ERROR 285008
-#define AMCONTINUOUSSCANACTIONCONTROLLER_FAILE_TO_WRITE_FILE 285009
+#define AMCONTINUOUSSCANACTIONCONTROLLER_COULD_NOT_ADD_DETECTOR			285000
+#define AMCONTINUOUSSCANACTIONCONTROLLER_FILE_ALREADY_EXISTS			285001
+#define AMCONTINUOUSSCANACTIONCONTROLLER_COULD_NOT_OPEN_FILE			285002
+#define AMCONTINUOUSSCANACTIONCONTROLLER_UNKNOWN_FILE_ERROR				285003
+#define AMCONTINUOUSSCANACTIONCONTROLLER_FAILE_TO_WRITE_FILE			285004
 
 /// This class is the base class for all continuous based scan controllers.
 class AMContinuousScanActionController : public AMScanActionController
@@ -74,30 +65,7 @@ protected:
 	/// Helper function to place all of the code into the datastore.  Subclasses should re-implement because continuous data has a lot of implementation specific details.
 	virtual void onAxisFinished();
 	/// Helper method that places data into the appropriate places after a data available message.  Passes the message.  Subclasses should re-implement.
-	virtual void fillDataMaps(AMAgnosticDataAPIDataAvailableMessage *message);
-
-	/// Helper function responsible principally responsible for gathering the requiredBufferNames_ and timesScales_. Returns false if an error occurs.
-	virtual bool generateAnalysisMetaInfo();
-
-	/// Helper function responsible for generating the scalerChannelIndexMap. Returns false if an error occurs.
-	virtual bool generateScalerMaps();
-
-	/// Helper function responsible for finding available motion start indices
-	virtual bool findStartMotionIndices();
-
-	/// Helper function responsible for generating the axis feedback values
-	virtual bool generateAxisFeedbackValues();
-
-	/// Helper function responsible for generating all of the interpolation vectors and parameters
-	virtual bool generateInterpolatedParameters();
-	/// Helper function responsible for generating all of the interpolated data and placing it in the vectors for the scaler channels
-	virtual bool generateInterpolatedScalerVectors();
-	/// Helper function responsible for generating all of the interpolated data and placing it in the vectors for the ampteks
-	virtual bool generateInterpolatedAmptekVectors();
-
-	/// Helper function to place the interpolated data in the data store
-	virtual bool placeInterpolatedDataInDataStore();
-
+	virtual void fillDataMaps(AMAgnosticDataAPIDataAvailableMessage *message);	
 	/// Helper function to do necessary cleanup on client data requests
 	virtual bool cleanupClientDataRequests();
 
@@ -112,64 +80,18 @@ protected:
 
 	/// Map of bufferNames to client data requests
 	QMap<QString, AMDSClientDataRequest*> clientDataRequestMap_;
-	/// Map of other meta data collected at run time by request
-	QMap<QString, double> metaDataMap_;
-	/// The axis order mapping.  The mapping is <scan axis name, desired index>.  As with any list, the index should be between 0 and scanRank()-1.
-	QMap<QString, int> axisOrderMap_;
-
 	/// The insertion index for placing data
 	AMnDIndex insertionIndex_;
-
+	/// The axis order mapping.  The mapping is <scan axis name, desired index>.  As with any list, the index should be between 0 and scanRank()-1.
+	QMap<QString, int> axisOrderMap_;
 	/// Holds the list of AMDS buffer names we expect to have
 	QList<QString> requiredBufferNames_;
 	/// Holds the list of time scales for polling detectors
 	QList<int> timeScales_;
-	/// The maximum value in the previous list, for quick lookup
-	int largestBaseTimeScale_;
-	/// Holds a list of scaler channel detectors is any were found
-	QList<CLSAMDSScalerChannelDetector*> scalerChannelDetectors_;
-	/// Holds a list of amptek detectors is any were found
-	QList<CLSAmptekSDD123DetectorNew*> amptekDetectors_;
-
-	/// Holds the total amount of scaler data received
-	int scalerTotalCount_;
-	/// Holds the mapping from scaler channel index to scaler channel detector name
-	QMap<int, QString> scalerChannelIndexMap_;
-
-	/// Calculated expected duration from the base time scale used
-	int expectedDurationScaledToBaseTimeScale_;
-	/// A map of detectors (by name) to the start motion index each is using
-	QMap<QString, int> detectorStartMotionIndexMap_;
-
-	/// A map of the scaler channel detectors (by name) to each rebased vector
-	QMap<QString, QVector<qint32> > scalerChannelRebaseVectors_;
-	/// One of the scaler vectors that might have our movement data
-	QVector<qint32> encoderUpVector_;
-	/// The other scaler vector that might have our movement data
-	QVector<qint32> encoderDownVector_;
-	/// Convenience location to quickly call for the start index of scaler motion
-	int scalerInitiateMovementIndex_;
-
-	/// Vector of the actual recorded axis values
-	QVector<double> axisFeedbackValues_;
-
-	/// The interpolation resolution step
-	double resolutionStep_;
-	/// The size of the interpolated vectors, calculated from data available and parameters
-	double interpolatedSize_;
 	/// Flag for if this scan increases or decreases (based on actual feedbacks)
 	bool isUpScan_;
-	/// The vector of interpolated axis values
-	QVector<double> interpolatedAxis_;
-	/// The vector of interpolated midpoints (for binning interpolation)
-	QVector<double> interpolatedMidpoints_;
-	/// The vector that maps the midpoints above to fractional indices
-	QVector<double> interpolatedMidpointsMappingIndices_;
 
-	/// The actual interpolated data for the scaler channels
-	QMap<QString, QVector<double> > interpolatedScalerChannelVectors_;
-	/// The actual interpolated data for the ampteks
-	QMap<QString, QVector<double> > interpolatedAmptekVectors_;
+
 };
 
 #endif // AMCONTINUOUSSCANACTIONCONTROLLER_H
